requirements.txt
python-docx
langchain
langchain-community
weaviate-client
sentence-transformers
FlagEmbedding